scraper/requirements.txt
async-generator==1.10
attrs==22.1.0
awscrt==0.14.0
beautifulsoup4==4.11.1
boto3==1.24.62
botocore==1.27.62
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.5
cryptography==37.0.4
dataclasses-json==0.5.7
h11==0.13.0
idna==3.3
iniconfig==1.1.1
jmespath==1.0.1
marshmallow==3.17.1
marshmallow-enum==1.5.1
mypy-extensions==0.4.3
outcome==1.2.0
packaging==21.3
peewee==3.15.2
pluggy==1.0.0
py==1.11.0
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.1.3
python-dateutil==2.8.2
python-dotenv==0.21.0
PyYAML==6.0
s3transfer==0.6.0
selenium==4.3.0
selenium-stealth==1.0.6
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
tomli==2.0.1
tqdm==4.64.1
trio==0.21.0
trio-websocket==0.9.2
typing-inspect==0.8.0
typing_extensions==4.3.0
urllib3==1.26.11
wsproto==1.1.0
